Add The Big Down.
This commit statistics the trend defore/after market drop/rise.

The three figures show the result in three stock index. The x-axis
means the trading days before or after the market unheaval. The
y-axis means the diff between index price on x-days and index
price on market drop/rise day. The formulas is:
  y = avg((day_x_price - unheaval_day_price) / unhaval_day_price)

As shown in the picture, there are some statistical significance.
1. There is an one-year-long period of market rise before market
   unheaval.
2. Closely before the market unheaval day, there is a days-long
   period of market price down. The market price down before market
   drop is more intense.
3. Drop's times is less than rise, but drop's range is greater.

Signed-off-by: zhoujingyue <zhoujingyue@ruijie.com.cn>